Split MultiAnalyzerTest into shared and app
